requirements.txt
numpy
numba
Pillow
pyglet<1.5
pyyaml
gymnasium
matplotlib
scipy

pandas
scikit-image
tqdm
h5py
hdf5plugin
omegaconf
hydra-core
torch
tensorboard